Moved UniformBufferObject struct to separate file vulkan_descriptors.h. Moved createUniformBuffers and updateDescriptorSet to main code example. These methods will be generalized later and added to vulkan_descriptors.cpp. Removed GLM includes from VulkanEngine interface. Moved uniformBuffers and uniformBuffersMemory to main code example. Set original cameraEye as a float vector of size 3 instead of glm::vec3
